+++ src/view/films-container-view.js
@@ -1,0 +1,5 @@
+export const createFilmsContainerTemplate = () => (
+  `<section class="films">
+
+  </section>`
+);
