+++ src/view/quantity-films-view.js
@@ -1,0 +1,3 @@
+export const createQuantityFilmsTemplate = (cards) => (
+  `<p>${ cards.length } movies inside</p>`
+);
